cast to long before multiply

+++ KeywordSearch/src/org/sleuthkit/autopsy/keywordsearch/SearchRunner.java
@@ -94,7 +94,7 @@ public final class SearchRunner {
         
         // start the timer, if needed
         if ((jobs.size() > 0) && (updateTimerRunning == false)) {
-            final long updateIntervalMs = KeywordSearchSettings.getUpdateFrequency().getTime() * 60 * 1000;
+            final long updateIntervalMs = ((long)KeywordSearchSettings.getUpdateFrequency().getTime()) * 60 * 1000;
             updateTimer.scheduleAtFixedRate(new UpdateTimerTask(), updateIntervalMs, updateIntervalMs);
             updateTimerRunning = true;
         }
